+++ base.h
@@ -4,7 +4,7 @@
 #include <sqlite3.h>
 
 int base_init(sqlite3** db, char* dbname);
-int base_save(sqlite3* db, double T, double H);
+int base_save(sqlite3* db, float T, float H);
 int base_load_last(sqlite3* db, unsigned int N, char** result);
 int base_load_between(sqlite3* db, const char* from, const char* to, char** result);
 int base_query (sqlite3* db, const char* q);
